refactor(email_api): extraire startup helpers (main 646→101 LOC)
